requirements.txt
aiogram==3.13.1
pandas==2.2.3
openpyxl==3.1.5
python-dotenv==1.0.1
aiosqlite==0.20.0
